修复不能跨设备移到文件错误

+++ services/file_extension_hap/entry/src/main/ets/FileExtensionAbility/FileExtensionAbility.ts
@@ -41,10 +41,10 @@ export default class FileExtAbility extends Extension {
     }
 
     checkUri(uri) {
-        if(uri.indexOf(FILE_ACCESS) == 0) {
+        if (uri.indexOf(FILE_ACCESS) == 0) {
             uri = uri.replace(FILE_ACCESS, '');
             return /^\/([^\/]+\/?)+$/.test(uri);
-        } else{
+        } else {
             return false;
         }
     }
@@ -59,13 +59,15 @@ export default class FileExtAbility extends Extension {
         if (arr[1].indexOf('/') > 0) {
             path = path.replace(arr[1].split('/')[0], '');
         }
+        if (path.charAt(path.length - 1) == '/') {
+            path = path.substr(0, path.length - 1);
+        }
         return path;
     }
 
     genNewFileUri(uri, displayName) {
-        let arr = uri.split('/');
         let newFileUri = uri;
-        if (arr.pop() == '') {
+        if (uri.charAt(uri.length - 1) == '/') {
             newFileUri += displayName;
         } else {
             newFileUri += '/' + displayName;
@@ -110,13 +112,14 @@ export default class FileExtAbility extends Extension {
             if (stat.isDirectory()) {
                 let dir = fileio.opendirSync(path);
                 let hasNextFile = true;
+                cb(path, true, hasNextFile);
                 while (hasNextFile) {
                     try {
                         let dirent = dir.readSync();
                         this.listDir(path + '/' + dirent.name, cb);
                     } catch (e) {
                         hasNextFile = false;
-                        cb(path, true);
+                        cb(path, true, hasNextFile);
                     }
                 }
             } else {
@@ -129,7 +132,7 @@ export default class FileExtAbility extends Extension {
     }
 
     openFile(sourceFileUri, flags) {
-        if(!this.checkUri(sourceFileUri)) {
+        if (!this.checkUri(sourceFileUri)) {
             return -1;
         }
         let fd = 0;
@@ -154,7 +157,7 @@ export default class FileExtAbility extends Extension {
     }
 
     createFile(parentUri, displayName) {
-        if(!this.checkUri(parentUri)) {
+        if (!this.checkUri(parentUri)) {
             return '';
         }
         try {
@@ -169,7 +172,7 @@ export default class FileExtAbility extends Extension {
     }
 
     mkdir(parentUri, displayName) {
-        if(!this.checkUri(parentUri)) {
+        if (!this.checkUri(parentUri)) {
             return '';
         }
         try {
@@ -184,15 +187,17 @@ export default class FileExtAbility extends Extension {
     }
 
     delete(selectFileUri) {
-        if(!this.checkUri(selectFileUri)) {
+        if (!this.checkUri(selectFileUri)) {
             return -1;
         }
         let path = this.getPath(selectFileUri);
         let code = 0;
-        this.listDir(path, function (filePath, isDirectory) {
+        this.listDir(path, function (filePath, isDirectory, hasNextFile) {
             try {
                 if (isDirectory) {
-                    fileio.rmdirSync(filePath);
+                    if (!hasNextFile) {
+                        fileio.rmdirSync(filePath);
+                    }
                 } else {
                     fileio.unlinkSync(filePath);
                 }
@@ -205,24 +210,46 @@ export default class FileExtAbility extends Extension {
     }
 
     move(sourceFileUri, targetParentUri) {
-        if(!this.checkUri(sourceFileUri) || !this.checkUri(targetParentUri)) {
+        if (!this.checkUri(sourceFileUri) || !this.checkUri(targetParentUri)) {
             return '';
         }
-        try {
-            let displayName = this.getFileName(sourceFileUri);
-            let newFileUri = this.genNewFileUri(targetParentUri, displayName);
-            let oldPath = this.getPath(sourceFileUri);
-            let newPath = this.getPath(newFileUri);
-            fileio.renameSync(oldPath, newPath);
+        let displayName = this.getFileName(sourceFileUri);
+        let newFileUri = this.genNewFileUri(targetParentUri, displayName);
+        let oldPath = this.getPath(sourceFileUri);
+        let newPath = this.getPath(newFileUri);
+        if (oldPath == newPath) {
             return newFileUri;
-        } catch (e) {
-            hilog.error(DOMAIN_CODE, TAG, 'move error ' + e.message);
+        } else if (newPath.indexOf(oldPath) == 0) {
             return '';
+        }
+        let hasError = false;
+        this.listDir(oldPath, function (filePath, isDirectory, hasNextFile) {
+            try {
+                let newFilePath = filePath.replace(oldPath, newPath);
+                if (isDirectory) {
+                    if (hasNextFile) {
+                        fileio.mkdirSync(newFilePath);
+                    } else {
+                        fileio.rmdirSync(filePath);
+                    }
+                } else {
+                    fileio.copyFileSync(filePath, newFilePath);
+                    fileio.unlinkSync(filePath);
+                }
+            } catch (e) {
+                hasError = true;
+                hilog.error(DOMAIN_CODE, TAG, 'move error ' + e.message);
+            }
+        });
+        if (hasError) {
+            return '';
+        } else {
+            return newFileUri;
         }
     }
 
     rename(sourceFileUri, displayName) {
-        if(!this.checkUri(sourceFileUri)) {
+        if (!this.checkUri(sourceFileUri)) {
             return '';
         }
         try {
@@ -238,7 +265,7 @@ export default class FileExtAbility extends Extension {
     }
 
     query(sourceFileUri) {
-        if(!this.checkUri(sourceFileUri)) {
+        if (!this.checkUri(sourceFileUri)) {
             return null;
         }
         try {
@@ -259,7 +286,7 @@ export default class FileExtAbility extends Extension {
     }
 
     listFile(sourceFileUri) {
-        if(!this.checkUri(sourceFileUri)) {
+        if (!this.checkUri(sourceFileUri)) {
             return [];
         }
         let infos = [];
@@ -280,7 +307,6 @@ export default class FileExtAbility extends Extension {
                         mimeType: '',
                     });
                 } catch (e) {
-                    hilog.error(DOMAIN_CODE, TAG, 'listFile error ' + e.message);
                     hasNextFile = false;
                 }
             }
